+++ perception/radar_object_tracker/include/autoware/radar_object_tracker/association/solver/gnn_solver.hpp
@@ -1,4 +1,4 @@
-// Copyright 2020 Tier IV, Inc.
+// Copyright 2021 Tier IV, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef RADAR_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
-#define RADAR_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
+#ifndef AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
+#define AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
 
-#include "model/constant_turn_rate_motion_tracker.hpp"
-#include "model/linear_motion_tracker.hpp"
-#include "model/tracker_base.hpp"
+#include "autoware/radar_object_tracker/association/solver/gnn_solver_interface.hpp"
+#include "autoware/radar_object_tracker/association/solver/mu_ssp.hpp"
+#include "autoware/radar_object_tracker/association/solver/ssp.hpp"
 
-#endif  // RADAR_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
+#endif  // AUTOWARE__RADAR_OBJECT_TRACKER__ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
