+++ perception/radar_object_tracker/include/autoware/radar_object_tracker/utils/utils.hpp
@@ -16,16 +16,16 @@
 // Author: v1.0 Yukihiro Saito
 //
 
-#ifndef RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
-#define RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
+#ifndef AUTOWARE__RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
+#define AUTOWARE__RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
 
 #include <Eigen/Core>
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
 
-#include <autoware_perception_msgs/msg/detected_object.hpp>
-#include <autoware_perception_msgs/msg/shape.hpp>
-#include <autoware_perception_msgs/msg/tracked_object.hpp>
+#include "autoware_perception_msgs/msg/detected_object.hpp"
+#include "autoware_perception_msgs/msg/shape.hpp"
+#include "autoware_perception_msgs/msg/tracked_object.hpp"
 #include <geometry_msgs/msg/polygon.hpp>
 #include <geometry_msgs/msg/vector3.hpp>
 
@@ -81,4 +81,4 @@ Eigen::MatrixXd stackMatricesDiagonally(const std::vector<Eigen::MatrixXd> & mat
 
 }  // namespace utils
 
-#endif  // RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
+#endif  // AUTOWARE__RADAR_OBJECT_TRACKER__UTILS__UTILS_HPP_
